Code cleanup and adding more tests
